// app-service/src/interface/enum.ts
export enum HttpStatusMessage {
  CONTINUE = "CONTINUE",
  SWITCHING_PROTOCOLS = "SWITCHING_PROTOCOLS",
  PROCESSING = "PROCESSING",
  OK = "OK",
  CREATED = "CREATED",
  ACCEPTED = "ACCEPTED",
  NON_AUTHORITATIVE_INFORMATION = "NON_AUTHORITATIVE_INFORMATION",
  NO_CONTENT = "NO_CONTENT",
  RESET_CONTENT = "RESET_CONTENT",
  PARTIAL_CONTENT = "PARTIAL_CONTENT",
  AMBIGUOUS = "AMBIGUOUS",
  MOVED_PERMANENTLY = "MOVED_PERMANENTLY",
  FOUND = "FOUND",
  SEE_OTHER = "SEE_OTHER",
  NOT_MODIFIED = "NOT_MODIFIED",
  TEMPORARY_REDIRECT = "TEMPORARY_REDIRECT",
  PERMANENT_REDIRECT = "PERMANENT_REDIRECT",
  BAD_REQUEST = "BAD_REQUEST",
  UNAUTHORIZED = "UNAUTHORIZED",
  PAYMENT_REQUIRED = "PAYMENT_REQUIRED",
  FORBIDDEN = "FORBIDDEN",
  NOT_FOUND = "NOT_FOUND",
  METHOD_NOT_ALLOWED = "METHOD_NOT_ALLOWED",
  NOT_ACCEPTABLE = "NOT_ACCEPTABLE",
  PROXY_AUTHENTICATION_REQUIRED = "PROXY_AUTHENTICATION_REQUIRED",
  REQUEST_TIMEOUT = "REQUEST_TIMEOUT",
  CONFLICT = "CONFLICT",
  GONE = "GONE",
  LENGTH_REQUIRED = "LENGTH_REQUIRED",
  PRECONDITION_FAILED = "PRECONDITION_FAILED",
  PAYLOAD_TOO_LARGE = "PAYLOAD_TOO_LARGE",
  URI_TOO_LONG = "URI_TOO_LONG",
  UNSUPPORTED_MEDIA_TYPE = "UNSUPPORTED_MEDIA_TYPE",
  REQUESTED_RANGE_NOT_SATISFIABLE = "REQUESTED_RANGE_NOT_SATISFIABLE",
  EXPECTATION_FAILED = "EXPECTATION_FAILED",
  I_AM_A_TEAPOT = "I_AM_A_TEAPOT",
  MISDIRECTED = "MISDIRECTED",
  UNPROCESSABLE_ENTITY = "UNPROCESSABLE_ENTITY",
  FAILED_DEPENDENCY = "FAILED_DEPENDENCY",
  PRECONDITION_REQUIRED = "PRECONDITION_REQUIRED",
  TOO_MANY_REQUESTS = "TOO_MANY_REQUESTS",
  INTERNAL_SERVER_ERROR = "INTERNAL_SERVER_ERROR",
  NOT_IMPLEMENTED = "NOT_IMPLEMENTED",
  BAD_GATEWAY = "BAD_GATEWAY",
  SERVICE_UNAVAILABLE = "SERVICE_UNAVAILABLE",
  GATEWAY_TIMEOUT = "GATEWAY_TIMEOUT",
  HTTP_VERSION_NOT_SUPPORTED = "HTTP_VERSION_NOT_SUPPORTED",
  ACCOUNT_DELETED = "ACCOUNT_DELETED",
  ACCOUNT_INACTIVE = "ACCOUNT_INACTIVE",
  ACCOUNT_LOCKED = "ACCOUNT_LOCKED",
  DEFAULT_MESSAGE = "DEFAULT_MESSAGE",
  INVALID_IP77_PASSWORD = "INVALID_IP77_PASSWORD",
}
export enum USER_TYPE {
  BUYER = "buyer",
  SELLER = "seller",
}
export enum SELLER {
  SELLER = "seller",
}
export enum COLLECTION {
  USER = "user",
  SESSION = "session",
  RESTAURANT = "restaurant",
  FOOD = "food",
  CATEGORY = "category",
  CART = "cart",
  ORDER = 'order'
}
export enum orderState {
  PENDING = "PENDING",
  SUCCESS = "SUCCESS",
}
export enum ExceptionMessage {
  DISTANCE_ERROR= 'Maximum distance reached',
  INVALID_ADDRESS="INVALID ADDRESS",
  NOT_FOUND = "NOT_FOUND",
  DEFAULT_MESSAGE = "DEFAULT_MESSAGE",
  EMAIL_ALREADY_EXIST = "EMAIL_ALREADY_EXIST",
  VERIFICATION_FAILED = "VERIFICATION_FAILED",
  EMAIL_NOT_EXISTS = "EMAIL_NOT_EXISTS",
  LOGIN_FAILED = "LOGIN_FAILED",
  SOMETHING_WENT_WRONG = "SOMETHING_WENT_WRONG",
  USER_NOT_EXISTS = "USER_NOT_EXISTS",
  USER_ALREADY_EXIST = "USER_ALREADY_EXIST",
  AUTH_INVALID_TOKEN = "AUTH_INVALID_TOKEN",
  AUTH_INVALID_STRATEGY = "AUTH_INVALID_STRATEGY",
  OTP_EXPIRED = "OTP_EXPIRED",
  SESSION_NOT_FOUND = "SESSION_NOT_FOUND",
  INVALID_REQUEST = "INVALID_REQUEST",
  INCORRECT_OTP = "INCORRECT_OTP",
  INVALID_APIKEY = "INVALID_APIKEY",
  INVALID_OTP = "INVALID_OTP",
  INVALID_PASSWORD = "INVALID_PASSWORD",
  UNAUTHORIZED = "UNAUTHORIZED",
  MOVED_PERMANENTLY = "MOVED_PERMANENTLY",
  USER_BLOCKED = "USER_BLOCKED",
  USER_NOT_FOUND = "USER_NOT_FOUND",
  TOKENIZATION_ERROR = "TOKENIZATION_ERROR",
  LINK_EXPIRE = "Link Expired",
  CONNECTION_NOT_FOUND = "CONNECTION_NOT_FOUND",
  FILE_ERROR = "FILE_NOT_FOUND",
}